Added LabJack timer configuration to LabJack.java
Added OutOfRange and ConfigurationError exceptions to Ramping.java, MotorController.java and MotorControllerApplication.java. These exceptions are raised when the MotorController tries to output a value and the LabJack module is either incorrectly configured or the given value is out of range.

Updated RudderControllerApplication.java to reflect the changes in RudderController.java